Web App/requirements.txt
requests
python-dotenv
python-barcode
pandas
Flask
pillow